requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==2.2
django-admin-interface==0.16.1
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-crispy-forms==1.11.1
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.5.0
django-js-asset==1.2.2
djangorestframework==3.12.2
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mysqlclient==2.0.3
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
odfpy==1.4.1
openpyxl==3.0.6
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pycparser==2.20
Pygments==2.8.0
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-barcode==0.13.1
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qrcode==6.1
qtconsole==5.0.2
QtPy==1.9.0
reportlab==3.5.59
segno==1.3.1
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
virtualenv==20.4.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
